Extracted some helper methods into a TestUtils class
